Removed SaveProjectsRule
