+++ collect_app/src/main/java/org/odk/collect/android/projects/ProjectsRepository.kt
@@ -8,4 +8,6 @@ interface ProjectsRepository {
     fun add(project: Project)
 
     fun delete(uuid: String)
+
+    fun deleteAll()
 }
